+++ opentelemetry-etw-logs/src/logs/mod.rs
@@ -6,5 +6,5 @@ mod reentrant_logprocessor;
 mod with_etw_exporter;
 
 pub use exporter_options::ExporterOptions;
-pub use reentrant_logprocessor::new_log_processor;
+pub use reentrant_logprocessor::etw_log_processor;
 pub use with_etw_exporter::ETWLoggerProviderBuilderExt;