+++ opentelemetry-etw-logs/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Added validation to provider name.
 - Added optional feature `serde_json` to serialize List and Maps.
-- Added `new_log_processor()` method that returns an `impl LogProcessor`.
+- Added `etw_log_processor()` method that returns an `impl LogProcessor`.
 - Added `ExporterOptions` which uses a builder pattern to configure the exporter.
 - The `with_etw_exporter` trait method now accepts an `ExporterOptions` instance: